+++ traffic_router/connector/src/main/java/org/apache/traffic_control/traffic_router/protocol/LanguidProtocol.java
@@ -17,10 +17,11 @@ package org.apache.traffic_control.traffic_router.protocol;
 
 
 import org.apache.coyote.http11.Http11NioProtocol;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 public class LanguidProtocol extends Http11NioProtocol implements RouterProtocolHandler {
-	private static final Logger log = Logger.getLogger(LanguidProtocol.class);
+	private static final Logger log = LogManager.getLogger(LanguidProtocol.class);
 	private boolean ready = false;
 	private boolean initialized = false;
 	private String mbeanPath;