+++ traffic_router/core/src/main/java/org/apache/traffic_control/traffic_router/core/loc/MaxmindGeolocationService.java
@@ -19,7 +19,8 @@ import java.io.File;
 import java.io.IOException;
 import java.net.InetAddress;
 
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import com.maxmind.geoip2.DatabaseReader;
 import com.maxmind.geoip2.exception.AddressNotFoundException;
@@ -30,7 +31,7 @@ import org.apache.traffic_control.traffic_router.geolocation.GeolocationExceptio
 import org.apache.traffic_control.traffic_router.geolocation.GeolocationService;
 
 public class MaxmindGeolocationService implements GeolocationService {
-	private static final Logger LOGGER = Logger.getLogger(MaxmindGeolocationService.class);
+	private static final Logger LOGGER = LogManager.getLogger(MaxmindGeolocationService.class);
 	private boolean initialized = false;
 	private DatabaseReader databaseReader;
 	private File databaseFile;