+++ traffic_router/core/src/test/java/org/apache/traffic_control/traffic_router/core/ds/DeliveryServiceMatcherTest.java
@@ -19,6 +19,7 @@ import org.apache.traffic_control.traffic_router.core.request.HTTPRequest;
 import org.apache.traffic_control.traffic_router.core.request.Request;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.powermock.core.classloader.annotations.PowerMockIgnore;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
@@ -33,6 +34,7 @@ import static org.mockito.Mockito.when;
 
 @PrepareForTest(DeliveryService.class)
 @RunWith(PowerMockRunner.class)
+@PowerMockIgnore("javax.management.*")
 public class DeliveryServiceMatcherTest {
 	@Test
 	public void itReturnsTrueWhenAllMatchersPass() {