+++ traffic_router/core/src/test/java/org/apache/traffic_control/traffic_router/core/loc/GeoTest.java
@@ -16,7 +16,8 @@
 package org.apache.traffic_control.traffic_router.core.loc;
 
 import org.apache.traffic_control.traffic_router.core.util.IntegrationTest;
-import org.apache.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -36,7 +37,7 @@ import static org.hamcrest.Matchers.equalTo;
 
 @Category(IntegrationTest.class)
 public class GeoTest {
-	private static final Logger LOGGER = Logger.getLogger(GeoTest.class);
+	private static final Logger LOGGER = LogManager.getLogger(GeoTest.class);
 
 	private GeolocationDatabaseUpdater geolocationDatabaseUpdater;
 	private MaxmindGeolocationService maxmindGeolocationService;