+++ traffic_router/core/src/test/java/org/apache/traffic_control/traffic_router/core/router/DNSRoutingMissesTest.java
@@ -31,6 +31,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.powermock.core.classloader.annotations.PowerMockIgnore;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import org.powermock.reflect.Whitebox;
@@ -44,6 +45,7 @@ import static org.powermock.reflect.Whitebox.setInternalState;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({DeliveryService.class, TrafficRouter.class})
+@PowerMockIgnore("javax.management.*")
 public class DNSRoutingMissesTest {
 
     private DNSRequest request;