+++ traffic_router/shared/src/main/java/org/apache/traffic_control/traffic_router/shared/SigningData.java
@@ -15,6 +15,8 @@
 
 package org.apache.traffic_control.traffic_router.shared;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.xbill.DNS.Record;
 import org.xbill.DNS.Section;
 
@@ -571,6 +573,8 @@ public class SigningData {
 	public static List<Record> signedList;
 	public static List<Record> postZoneList;
 	public static List<Record> dsRecordList = new ArrayList<>();
+
+	private static final Logger LOGGER = LogManager.getLogger(SigningData.class);
 
 	public static List<String> getStringsNamedLike(String name, Class clazz) {
 		return Arrays.asList(clazz.getDeclaredFields()).stream()